requirements.txt
Flask
flask-cors
tensorflow
keras-cv
Pillow
numpy
